move load translations into utils
